refactor(meta): extract file_format.proto and stage.proto.
