+++ src/meta/proto-conv/src/lib.rs
@@ -63,9 +63,11 @@
 mod config_from_to_protobuf_impl;
 mod database_from_to_protobuf_impl;
 mod datetime_from_to_protobuf_impl;
+mod file_format_from_to_protobuf_impl;
 mod from_to_protobuf;
 mod schema_from_to_protobuf_impl;
 mod share_from_to_protobuf_impl;
+mod stage_from_to_protobuf_impl;
 mod table_from_to_protobuf_impl;
 mod user_from_to_protobuf_impl;
 mod util;
