+++ src/meta/protos/proto/config.proto
@@ -16,6 +16,16 @@ syntax = "proto3";
 
 package databend_proto;
 
+message StorageConfig {
+  oneof storage {
+    S3StorageConfig s3 = 1;
+    FsStorageConfig fs = 2;
+    GcsStorageConfig gcs = 3;
+    OssStorageConfig oss = 4;
+    WebhdfsStorageConfig webhdfs = 5;
+  }
+}
+
 message S3StorageConfig {
   uint64 version = 100;
   uint64 min_reader_ver = 101;
