+++ src/meta/protos/proto/stage.proto
@@ -1,0 +1,86 @@
+// Copyright 2023 Datafuse Labs.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+syntax = "proto3";
+package databend_proto;
+
+import "config.proto";
+import "user.proto";
+import "file_format.proto";
+import "datatype.proto";
+
+message StageInfo {
+  uint64 ver = 100;
+  uint64 min_reader_ver = 101;
+
+  enum StageType {
+    LegacyInternal = 0;
+    External = 1;
+    Internal = 2;
+    User = 3;
+  }
+
+  message StageParams { StorageConfig storage = 1; }
+
+  message OnErrorMode {
+    oneof mode {
+      Empty None = 1;
+      Empty Continue = 2;
+      Empty SkipFile = 3;
+      uint64 SkipFileNum = 4;
+      Empty AbortStatement = 5;
+      uint64 AbortNum = 6;
+    }
+  }
+
+  message CopyOptions {
+    OnErrorMode on_error = 1;
+    uint64 size_limit = 2;
+    uint64 max_files = 10;
+    bool purge = 3;
+    bool single = 4;
+    uint64 max_file_size = 5;
+    uint64 split_size = 6;
+  }
+
+
+  string stage_name = 1;
+
+  StageType stage_type = 2;
+
+  StageParams stage_params = 3;
+
+  FileFormatOptions file_format_options = 4;
+
+  CopyOptions copy_options = 5;
+
+  string comment = 6;
+
+  uint64 number_of_files = 7;
+
+  optional UserIdentity creator = 8;
+}
+
+message StageFile {
+  uint64 ver = 100;
+  uint64 min_reader_ver = 101;
+
+  string path = 1;
+  uint64 size = 2;
+  optional string md5 = 3;
+  string last_modified = 4;
+  optional UserIdentity creator = 5;
+  optional string etag = 6;
+}
+
