+++ src/meta/protos/proto/user.proto
@@ -16,8 +16,6 @@ syntax = "proto3";
 
 package databend_proto;
 
-import "datatype.proto";
-import "config.proto";
 
 message AuthInfo {
   uint64 ver = 100;
@@ -118,135 +116,3 @@ message UserIdentity {
   string username = 1;
   string hostname = 2;
 }
-
-message StageInfo {
-  uint64 ver = 100;
-  uint64 min_reader_ver = 101;
-
-  enum StageType {
-    LegacyInternal = 0;
-    External = 1;
-    Internal = 2;
-    User = 3;
-  }
-
-  message StageStorage {
-    oneof storage {
-      S3StorageConfig s3 = 1;
-      FsStorageConfig fs = 2;
-      GcsStorageConfig gcs = 3;
-      OssStorageConfig oss = 4;
-      WebhdfsStorageConfig webhdfs = 5;
-    }
-  }
-
-  message StageParams { StageStorage storage = 1; }
-
-  enum StageFileFormatType {
-    Csv = 0;
-    Json = 1;
-    Avro = 2;
-    Orc = 3;
-    Parquet = 4;
-    Xml = 5;
-    NdJson = 6;
-    Tsv = 7;
-  }
-
-  enum StageFileCompression {
-    Auto = 0;
-    Gzip = 1;
-    Bz2 = 2;
-    Brotli = 3;
-    Zstd = 4;
-    Deflate = 5;
-    RawDeflate = 6;
-    Lzo = 7;
-    Snappy = 8;
-    // We used to assign `None` as 9, it's a mistake.
-    // Please change this value to 0 instead in next version bump.
-    None = 9;
-    Xz = 10;
-  }
-
-  message FileFormatOptions {
-    uint64 ver = 100;
-    uint64 min_reader_ver = 101;
-
-    StageFileFormatType format = 1;
-    // Number of lines at the start of the file to skip.
-    uint64 skip_header = 2;
-
-    string field_delimiter = 3;
-
-    string record_delimiter = 4;
-
-    StageFileCompression compression = 5;
-
-    string escape = 6;
-
-    string row_tag = 7;
-
-    string nan_display = 8;
-
-    string quote = 9;
-  }
-
-  message OnErrorMode {
-    oneof mode {
-      Empty None = 1;
-      Empty Continue = 2;
-      Empty SkipFile = 3;
-      uint64 SkipFileNum = 4;
-      Empty AbortStatement = 5;
-      uint64 AbortNum = 6;
-    }
-  }
-
-  message CopyOptions {
-    OnErrorMode on_error = 1;
-    uint64 size_limit = 2;
-    uint64 max_files = 10;
-    bool purge = 3;
-    bool single = 4;
-    uint64 max_file_size = 5;
-    uint64 split_size = 6;
-  }
-
-  message UserDefinedFileFormat {
-    uint64 ver = 100;
-    uint64 min_reader_ver = 101;
-    string name = 1;
-    FileFormatOptions file_format_options = 2;
-    UserIdentity creator = 3;
-  }
-
-  string stage_name = 1;
-
-  StageType stage_type = 2;
-
-  StageParams stage_params = 3;
-
-  FileFormatOptions file_format_options = 4;
-
-  CopyOptions copy_options = 5;
-
-  string comment = 6;
-
-  uint64 number_of_files = 7;
-
-  optional UserIdentity creator = 8;
-}
-
-message StageFile {
-  uint64 ver = 100;
-  uint64 min_reader_ver = 101;
-
-  string path = 1;
-  uint64 size = 2;
-  optional string md5 = 3;
-  string last_modified = 4;
-  optional UserIdentity creator = 5;
-  optional string etag = 6;
-}
-
